+++ src/QNEthernet.h
@@ -12,6 +12,9 @@
 #include <cstdint>
 #include <functional>
 
+#ifdef ARDUINO_ARCH_STM32
+#include <Arduino.h>  // STM32's Arduino needs this for namespace arduino
+#endif  // ARDUINO_ARCH_STM32
 #include <IPAddress.h>
 
 #include "lwip/apps/mdns_opts.h"