+++ src/qnethernet/QNDNSClient.h
@@ -16,6 +16,9 @@
 #include <functional>
 #include <limits>
 
+#ifdef ARDUINO_ARCH_STM32
+#include <Arduino.h>  // STM32's Arduino needs this for namespace arduino
+#endif  // ARDUINO_ARCH_STM32
 #include <IPAddress.h>
 
 #include "lwip/dns.h"