+++ src/qnethernet/QNEthernetFrame.h
@@ -15,6 +15,9 @@
 #include <cstdint>
 #include <vector>
 
+#ifdef ARDUINO_ARCH_STM32
+#include <Arduino.h>  // STM32's Arduino needs this for namespace arduino
+#endif  // ARDUINO_ARCH_STM32
 #include <Stream.h>
 
 #include "lwip/err.h"