+++ src/qnethernet/QNEthernetServer.h
@@ -14,6 +14,9 @@
 #include <cstddef>
 #include <cstdint>
 
+#ifdef ARDUINO_ARCH_STM32
+#include <Arduino.h>  // STM32's Arduino needs this for namespace arduino
+#endif  // ARDUINO_ARCH_STM32
 #include <Print.h>
 #include <Server.h>