+++ src/qnethernet/internal/ClientEx.h
@@ -9,6 +9,9 @@
 // C++ includes
 #include <cstdint>
 
+#ifdef ARDUINO_ARCH_STM32
+#include <Arduino.h>  // STM32's Arduino needs this for namespace arduino
+#endif  // ARDUINO_ARCH_STM32
 #include <Client.h>
 #include <IPAddress.h>