+++ src/qnethernet/util/PrintUtils.h
@@ -13,6 +13,9 @@
 #include <functional>
 #include <limits>
 
+#ifdef ARDUINO_ARCH_STM32
+#include <Arduino.h>  // STM32's Arduino needs this for namespace arduino
+#endif  // ARDUINO_ARCH_STM32
 #include <Print.h>
 #include <Stream.h>